+++ src/mplchart/primitives/barplot.py
@@ -1,6 +1,6 @@
 """BarPlot primitive"""
 
-from ..model import BindingPrimitive
+from ..model.primitive import BindingPrimitive
 from ..utils import get_label, series_data